requirements.txt
pandas==1.5.0
python_dotenv==1.0.0